pyproject.toml
[project]
name = "amplifier-module-tool-sage"
version = "0.1.0"
description = "Strategic advisor tool for architecture, design, product, and outcome-focused decisions via Gemini and OpenAI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "amplifier-core",
    "google-generativeai>=0.8.0",
    "openai>=1.0.0",
]

[project.entry-points."amplifier.modules"]
tool-sage = "amplifier_module_tool_sage"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources.amplifier-core]
git = "https://github.com/microsoft/amplifier-core"
branch = "main"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
